update db_version regardless of outcome
