+++ core/rs/core/src/c.rs
@@ -54,10 +54,6 @@ pub struct crsql_ExtData {
     pub pragmaDataVersion: ::core::ffi::c_int,
     pub dbVersion: sqlite::int64,
     pub pendingDbVersion: sqlite::int64,
-    // pub siteVersion: sqlite::int64,
-    // pub pendingSiteVersion: sqlite::int64,
-    // pub pSetSiteVersionStmt: *mut sqlite::stmt,
-    // pub nextSiteVersionSet: ::core::ffi::c_int,
     pub pragmaSchemaVersion: ::core::ffi::c_int,
     pub updatedTableInfosThisTx: ::core::ffi::c_int,
     pub pragmaSchemaVersionForTableInfos: ::core::ffi::c_int,
@@ -230,16 +226,6 @@ fn bindgen_test_layout_crsql_Changes_cursor() {
             stringify!(dbVersion)
         )
     );
-    // assert_eq!(
-    //     unsafe { ::core::ptr::addr_of!((*ptr).rowType) as usize - ptr as usize },
-    //     40usize,
-    //     // concat!(
-    //     //     "Offset of field: ",
-    //     //     stringify!(crsql_Changes_cursor),
-    //     //     "::",
-    //     //     stringify!(siteVersion)
-    //     // )
-    // );
     assert_eq!(
         unsafe { ::core::ptr::addr_of!((*ptr).rowType) as usize - ptr as usize },
         40usize,
@@ -337,46 +323,6 @@ fn bindgen_test_layout_crsql_ExtData() {
             stringify!(pendingDbVersion)
         )
     );
-    // assert_eq!(
-    //     unsafe { ::core::ptr::addr_of!((*ptr).siteVersion) as usize - ptr as usize },
-    //     40usize,
-    //     concat!(
-    //         "Offset of field: ",
-    //         stringify!(crsql_ExtData),
-    //         "::",
-    //         stringify!(siteVersion)
-    //     )
-    // );
-    // assert_eq!(
-    //     unsafe { ::core::ptr::addr_of!((*ptr).pendingSiteVersion) as usize - ptr as usize },
-    //     48usize,
-    //     concat!(
-    //         "Offset of field: ",
-    //         stringify!(crsql_ExtData),
-    //         "::",
-    //         stringify!(pendingSiteVersion)
-    //     )
-    // );
-    // assert_eq!(
-    //     unsafe { ::core::ptr::addr_of!((*ptr).pSetSiteVersionStmt) as usize - ptr as usize },
-    //     56usize,
-    //     concat!(
-    //         "Offset of field: ",
-    //         stringify!(crsql_ExtData),
-    //         "::",
-    //         stringify!(pSetSiteVersionStmt)
-    //     )
-    // );
-    // assert_eq!(
-    //     unsafe { ::core::ptr::addr_of!((*ptr).nextSiteVersionSet) as usize - ptr as usize },
-    //     64usize,
-    //     concat!(
-    //         "Offset of field: ",
-    //         stringify!(crsql_ExtData),
-    //         "::",
-    //         stringify!(nextSiteVersionSet)
-    //     )
-    // );
     assert_eq!(
         unsafe { ::core::ptr::addr_of!((*ptr).pragmaSchemaVersion) as usize - ptr as usize },
         40usize,
@@ -449,16 +395,6 @@ fn bindgen_test_layout_crsql_ExtData() {
             stringify!(lastDbVersions)
         )
     );
-    // assert_eq!(
-    //     unsafe { ::core::ptr::addr_of!((*ptr).pSiteVersionStmt) as usize - ptr as usize },
-    //     96usize,
-    //     concat!(
-    //         "Offset of field: ",
-    //         stringify!(crsql_ExtData),
-    //         "::",
-    //         stringify!(pSiteVersionStmt)
-    //     )
-    // );
     assert_eq!(
         unsafe { ::core::ptr::addr_of!((*ptr).tableInfos) as usize - ptr as usize },
         88usize,
@@ -469,16 +405,6 @@ fn bindgen_test_layout_crsql_ExtData() {
             stringify!(tableInfos)
         )
     );
-    // assert_eq!(
-    //     unsafe { ::core::ptr::addr_of!((*ptr).lastSiteVersions) as usize - ptr as usize },
-    //     112usize,
-    //     concat!(
-    //         "Offset of field: ",
-    //         stringify!(crsql_ExtData),
-    //         "::",
-    //         stringify!(lastSiteVersions)
-    //     )
-    // );
     assert_eq!(
         unsafe { ::core::ptr::addr_of!((*ptr).rowsImpacted) as usize - ptr as usize },
         96usize,
