+++ core/rs/core/src/changes_vtab.rs
@@ -41,7 +41,6 @@ fn changes_crsr_finalize(crsr: *mut crsql_Changes_cursor) -> c_int {
         }
         (*crsr).pRowStmt = null_mut();
         (*crsr).dbVersion = crate::consts::MIN_POSSIBLE_DB_VERSION;
-        // (*crsr).siteVersion = crate::consts::MIN_POSSIBLE_SITE_VERSION;
 
         rc
     }
@@ -209,7 +208,6 @@ fn get_clock_table_col_name(col: &Option<CrsqlChangesColumn>) -> Option<String> 
         Some(CrsqlChangesColumn::SiteId) => Some("site_id".to_string()),
         Some(CrsqlChangesColumn::Seq) => Some("seq".to_string()),
         Some(CrsqlChangesColumn::Cl) => Some("cl".to_string()),
-        // Some(CrsqlChangesColumn::SiteVrsn) => Some("site_vrsn".to_string()),
         None => None,
     }
 }
@@ -324,22 +322,17 @@ unsafe fn changes_next(
     cursor: *mut crsql_Changes_cursor,
     vtab: *mut sqlite::vtab,
 ) -> Result<ResultCode, ResultCode> {
-    // libc_print::libc_println!("chnages_next");
     if (*cursor).pChangesStmt.is_null() {
         let err = CString::new("pChangesStmt is null in changes_next")?;
         (*vtab).zErrMsg = err.into_raw();
         return Err(ResultCode::ABORT);
     }
 
-    // libc_print::libc_println!("chnages_next 2");
-
     if !(*cursor).pRowStmt.is_null() {
         let rc = reset_cached_stmt((*cursor).pRowStmt);
         (*cursor).pRowStmt = null_mut();
         rc?;
     }
-
-    // libc_print::libc_println!("chnages_next 3");
 
     let rc = (*cursor).pChangesStmt.step()?;
     if rc == ResultCode::DONE {
@@ -350,7 +343,6 @@ unsafe fn changes_next(
             return Err(ResultCode::ERROR);
         }
     }
-    // libc_print::libc_println!("chnages_next 4");
 
     // we had a row... we can do the rest
     let tbl = (*cursor)
@@ -368,17 +360,7 @@ unsafe fn changes_next(
     let changes_rowid = (*cursor)
         .pChangesStmt
         .column_int64(ClockUnionColumn::RowId as i32);
-    // let site_version = (*cursor)
-    //     .pChangesStmt
-    //     .column_int64(ClockUnionColumn::SiteVrsn as i32);
     (*cursor).dbVersion = db_version;
-    // (*cursor).siteVersion = site_version;
-
-    // libc_print::libc_println!(
-    //     "db_version = {}, site_version = {}",
-    //     db_version,
-    //     site_version
-    // );
 
     let tbl_infos = mem::ManuallyDrop::new(Box::from_raw(
         (*(*(*cursor).pTab).pExtData).tableInfos as *mut Vec<TableInfo>,
@@ -466,7 +448,6 @@ fn column_impl(
     let column = CrsqlChangesColumn::from_i32(i);
     // TODO: only de-reference where needed?
     let changes_stmt = unsafe { (*cursor).pChangesStmt };
-    // libc_print::libc_println!("getting column value for i = {} => {:?}", i, column);
     match column {
         Some(CrsqlChangesColumn::Tbl) => {
             ctx.result_value(changes_stmt.column_value(ClockUnionColumn::Tbl as i32));
@@ -513,9 +494,6 @@ fn column_impl(
         Some(CrsqlChangesColumn::Cl) => {
             ctx.result_value(changes_stmt.column_value(ClockUnionColumn::Cl as i32))
         }
-        // Some(CrsqlChangesColumn::SiteVrsn) => {
-        //     ctx.result_value(changes_stmt.column_value(ClockUnionColumn::SiteVrsn as i32))
-        // }
         None => return Err(ResultCode::MISUSE),
     }
 
@@ -544,7 +522,6 @@ pub extern "C" fn crsql_changes_update(
     argv: *mut *mut sqlite::value,
     row_id: *mut sqlite::int64,
 ) -> c_int {
-    // libc_print::libc_println!("crsql_changes_update");
     let args = sqlite::args!(argc, argv);
     let arg = args[0];
     if args.len() > 1 && arg.value_type() == ColumnType::Null {
