+++ core/rs/core/src/local_writes/after_delete.rs
@@ -39,10 +39,7 @@ fn after_delete(
     tbl_info: &TableInfo,
     pks_old: &[*mut value],
 ) -> Result<ResultCode, String> {
-    // libc_print::libc_println!("after delete");
     let db_version = crate::db_version::next_db_version(db, ext_data)?;
-    // let site_version = crate::site_version::next_site_version(db, ext_data)?;
-    // libc_print::libc_println!("next site version: {}", site_version);
     let seq = bump_seq(ext_data);
     let key = tbl_info
         .get_or_create_key_via_raw_values(db, pks_old)
