+++ core/rs/core/src/local_writes/after_insert.rs
@@ -39,12 +39,10 @@ fn after_insert(
     tbl_info: &TableInfo,
     pks_new: &[*mut value],
 ) -> Result<ResultCode, String> {
-    // libc_print::libc_println!("after_insert");
     let db_version = crate::db_version::next_db_version(db, ext_data)?;
-    // libc_print::libc_println!("next site version: {}", site_version);
     let (create_record_existed, key_new) = tbl_info
         .get_or_create_key_for_insert(db, pks_new)
-        .map_err(|_| "failed geteting or creating lookaside key")?;
+        .map_err(|_| "failed getting or creating lookaside key")?;
     if tbl_info.non_pks.is_empty() {
         let seq = bump_seq(ext_data);
         // just a sentinel record
@@ -54,12 +52,6 @@ fn after_insert(
         let seq = bump_seq(ext_data);
         update_create_record(db, tbl_info, key_new, db_version, seq)?;
     }
-
-    // now for each non-pk column, create or update the column record
-    // for col in tbl_info.non_pks.iter() {
-    //     let seq = bump_seq(ext_data);
-    //     super::mark_locally_updated(db, tbl_info, key_new, col, db_version, seq, site_version)?;
-    // }
 
     super::mark_locally_inserted(db, ext_data, tbl_info, key_new, db_version)?;
