+++ core/src/ext-data.h
@@ -21,11 +21,6 @@ struct crsql_ExtData {
   // if that transaction were to commit at the time this value is checked.
   sqlite3_int64 pendingDbVersion;
 
-  // sqlite3_int64 siteVersion;
-  // sqlite3_int64 pendingSiteVersion;
-  // sqlite3_stmt *pSetSiteVersionStmt;
-  // int nextSiteVersionSet;
-
   int pragmaSchemaVersion;
   int updatedTableInfosThisTx;
